+++ mullvad-relay-selector/src/relay_selector/mod.rs
@@ -80,6 +80,9 @@ pub static WIREGUARD_RETRY_ORDER: LazyLock<Vec<RelayQuery>> = LazyLock::new(|| {
             .udp2tcp()
             .ip_version(IpVersion::V6)
             .build(),
+        // 7
+        #[cfg(not(target_os = "android"))]
+        RelayQueryBuilder::wireguard().lwo().build(),
     ]
 });