+++ CHANGELOG.md
@@ -22,6 +22,8 @@ Line wrap the file at 100 chars.                                              Th
 * **Security**: in case of vulnerabilities.
 
 ## [Unreleased]
+### Added
+- Add new LWO obfuscation method for WireGuard.
 
 
 ## [2025.10-beta1] - 2025-09-16